+++ app/models.py
@@ -38,7 +38,7 @@ class Information(db.Model, UserMixin):
     Street = db.Column(db.String(length=30))
     Neighborhood = db.Column(db.String(length=30))
     Category = db.Column(db.String(length=3), nullable=False)
-    Cellphone = db.column(db.String(length=16))
+    Cellphone = db.Column(db.String(length=16))
     Birthday = db.Column(db.DateTime())
     # Criando A relação de 1 para 1 no banco de dados
     Account_id = db.Column(db.Integer, db.ForeignKey('account.Id'), nullable=False)
@@ -52,7 +52,7 @@ class Employees(db.Model, UserMixin):
     MName = db.Column(db.String(length=40))
     LName = db.Column(db.String(length=30), nullable=False)
     Email = db.Column(db.String(length=50))
-    Cellphone = db.column(db.String(length=16))
+    Cellphone = db.Column(db.String(length=16))
     Birthday = db.Column(db.DateTime())
     def get_id(self):
         return int(self.Id)